+++ src/main/java/me/flotsam/frettler/command/LookupCommand.java
@@ -39,14 +39,6 @@ public class LookupCommand implements Runnable {
   @Parameters(index = "0", description = "The chord notes to analyse", split = ",")
   Note[] notes = new Note[] {};
 
-  @Option(names = {"-v", "--verbose"},
-      description = "if you want some background to Frettlers application of music theory")
-  boolean verbose = false;
-
-  @Option(names = {"-m", "--mono"}, description = "Display in 'monochrome'")
-  @Getter
-  protected boolean mono;
-
   @Option(names = {"-r", "--rule"}, defaultValue="STRICT", description = "Chord lookup can be strict, relaxed or loose")
   @Getter
   protected Rule rule;