+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 28th April 2021
+- Added left handed instrument support
+
 # 3rd February 2021
 - Added the progression subcommand